Rename canvas to render2d as it no longer only does canvas
