+++ src/client/components/dashboard/dashboard_robot/view_model.ts
@@ -1,16 +1,16 @@
 import { createTransformer } from 'mobx'
 import { computed } from 'mobx'
 
-import { BasicAppearance } from '../../../canvas/appearance/basic_appearance'
-import { LineAppearance } from '../../../canvas/appearance/line_appearance'
-import { ArcGeometry } from '../../../canvas/geometry/arc_geometry'
-import { ArrowGeometry } from '../../../canvas/geometry/arrow_geometry'
-import { CircleGeometry } from '../../../canvas/geometry/circle_geometry'
-import { LineGeometry } from '../../../canvas/geometry/line_geometry'
-import { MarkerGeometry } from '../../../canvas/geometry/marker_geometry'
-import { TextGeometry } from '../../../canvas/geometry/text_geometry'
-import { Group } from '../../../canvas/object/group'
-import { Shape } from '../../../canvas/object/shape'
+import { BasicAppearance } from '../../../render2d/appearance/basic_appearance'
+import { LineAppearance } from '../../../render2d/appearance/line_appearance'
+import { ArcGeometry } from '../../../render2d/geometry/arc_geometry'
+import { ArrowGeometry } from '../../../render2d/geometry/arrow_geometry'
+import { CircleGeometry } from '../../../render2d/geometry/circle_geometry'
+import { LineGeometry } from '../../../render2d/geometry/line_geometry'
+import { MarkerGeometry } from '../../../render2d/geometry/marker_geometry'
+import { TextGeometry } from '../../../render2d/geometry/text_geometry'
+import { Group } from '../../../render2d/object/group'
+import { Shape } from '../../../render2d/object/shape'
 import { Transform } from '../../../math/transform'
 import { Vector2 } from '../../../math/vector2'
