+++ src/client/render2d/canvas_renderer.ts
