+++ src/client/render2d/geometry/arc_geometry.ts
