+++ src/client/render2d/geometry/arrow_geometry.ts
