+++ src/client/render2d/geometry/circle_geometry.ts
