+++ src/client/render2d/geometry/line_geometry.ts
