+++ src/client/render2d/geometry/marker_geometry.ts
