+++ src/client/render2d/geometry/polygon_geometry.ts
