+++ src/client/render2d/geometry/text_geometry.ts
