+++ src/client/render2d/object/group.ts
