+++ src/client/render2d/object/shape.ts
