+++ src/client/render2d/svg/arc.tsx
