+++ src/client/render2d/svg/arrow.tsx
