+++ src/client/render2d/svg/circle.tsx
