+++ src/client/render2d/svg/group.tsx
