+++ src/client/render2d/svg/line.tsx
