+++ src/client/render2d/svg/marker.tsx
