+++ src/client/render2d/svg/polygon.tsx
