+++ src/client/render2d/svg/svg.tsx
