+++ src/client/render2d/svg/tests/arc.tests.tsx
