+++ src/client/render2d/svg/tests/group.tests.tsx
