+++ src/client/render2d/svg/tests/polygon.tests.tsx
