+++ src/client/render2d/svg/tests/text.tests.tsx
