+++ src/client/render2d/svg/text.tsx
