+++ src/client/render2d/svg_renderer.tsx
